+++ pkg/enterprise/generator/basic_auth/basic_auth_test.go
@@ -16,7 +16,7 @@
 
 // Copyright External Secrets Inc. All Rights Reserved
 
-package basic_auth
+package basicauth
 
 import (
 	"errors"